+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 
 - **Appalerts**
     - Display project badge in alert (#790, #801)
+    - Dismiss all link in title bar badge (#802)
 - **Projectroles**
     - ``exact`` kwarg for ``assert_element_count()`` in UI tests (#798)
     - Custom email header and footer (#789)